fix: 🐛 handle lgtm alerts (#143)
* style: 🎨 code style issues with Prettier

* fix: 🐛 update register link for bio.tools

* refactor: ♻️ update help text for programming language

* refactor: ♻️ add text to show scopes for github

* refactor: ♻️ update wording for github repo and zenodo repo

* build: 📦️ update commit message for actions

* feat: ✨ add method to upgrade old datasets for new app versions (#141)

* style: 🎨 code style issues with Prettier

* feat: ✨ add method to upgrade old datasets for new app versions

* fix: 🐛 update show all projects for new version


* build: 📦️ align node versions

* fix: 🐛 handle lgtm alerts (#142)

* refactor: ♻️ pretty print generated json file

* style: 🎨 fix code style issues with Prettier

* revert: auth window

* fix: 🐛 update incorrect metadata key in codemeta

* style: 🎨 fix code style issues with Prettier

* chore: 🔧 sync with v1.3.1

* fix: 🐛 handle lgtm alerts
