+++ dev/linux-fairshare-dev.yml
@@ -16,7 +16,7 @@ dependencies:
   - libzlib=1.2.11
   - llvm-openmp=12.0.1
   - ncurses=6.3
-  - nodejs=16.14.0
+  - nodejs=16.14.2
   - openssl=1.1.1l
   - pip=21.2.4
   - python=3.9.7
