+++ dev/mac-fairshare-release.yml
@@ -11,7 +11,7 @@ dependencies:
   - libuv=1.42.0
   - libzlib=1.2.11
   - ncurses=6.3
-  - nodejs=16.14.0
+  - nodejs=16.14.2
   - openssl=1.1.1l
   - pip=21.2.4
   - python=3.9.7
