+++ src/pyflask/utilities.py
@@ -139,10 +139,9 @@ def fileExistInFolder(folder_path, file_name):
     try:
         for filename in os.listdir(folder_path):
             if filename == file_name:
-                import json
-
                 fileFullName = os.path.join(folder_path, filename)
                 f = open(fileFullName)
+
                 return json.load(f)
         return "Not Found"
     except Exception as e:
