+++ src/views/workflow/Code/CreateMetadata.vue
@@ -864,6 +864,10 @@
                               popoverContent="All programming languages used in this software. Select from the suggested list or type your own."
                             ></form-help-content>
                           </div>
+                          <p class="pt-2 text-xs text-zinc-600">
+                            Press <code>'Enter'</code> after typing to add a programming language
+                            that is not in the list.
+                          </p>
                         </el-form-item>
 
                         <el-form-item label="Runtime platform">
@@ -1699,7 +1703,7 @@ export default {
       });
     },
     initializeEmptyObjects(root, obj) {
-      if (typeof obj === "undefined") {
+      if (obj === undefined) {
         root[obj] = {};
       }
     },
@@ -2053,7 +2057,7 @@ export default {
       let authors = [];
 
       for (const contributor of contributors) {
-        response = await axios
+        await axios
           .get(`${process.env.VUE_APP_GITHUB_SERVER_URL}/users/${contributor}`, {
             params: {
               accept: "application/vnd.github.v3+json",
