+++ src/views/zenodo/ZenodoMetadata.vue
@@ -1162,7 +1162,7 @@ export default {
       });
     },
     initializeEmptyObjects(root, obj) {
-      if (typeof obj === "undefined") {
+      if (obj === undefined) {
         root[obj] = {};
       }
     },
